% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hero_section.R
\name{create_hero_section}
\alias{create_hero_section}
\title{Create hero section of the homepage}
\usage{
create_hero_section(title, subtitle)
}
\arguments{
\item{title}{Title of the app}

\item{subtitle}{Subtitle of the app}
}
\value{
Div.
}
\description{
Create hero section of the homepage
}
