% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_slides.R
\name{create_slides}
\alias{create_slides}
\title{Create revealjs slides using quarto}
\usage{
create_slides(input_qmd, theme, title, prompt, story, images, audios)
}
\arguments{
\item{input_qmd}{Path to qmd file}

\item{theme}{Revealjs theme}

\item{title}{Story title}

\item{prompt}{Prompt for creating story}

\item{story}{Story text}

\item{images}{List of binary images}

\item{audios}{List of binary audios}
}
\value{
Quarto revealjs slides
}
\description{
Create revealjs slides using quarto
}
