% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_prompts.R
\name{generate_prompts}
\alias{generate_prompts}
\title{Get prompt and negative prompt based on style}
\usage{
generate_prompts(response_prompts, style)
}
\arguments{
\item{response_prompts}{Vector of prompts for generating images}

\item{style}{A character string specifying the desired style. Choices are:
"anime", "comics", "lego_movie", "play_doh", "ethereal_fantasy", "isometric",
"line_art", "origami", "pixel_art", "abstract", "impressionist",
"renaissance", "watercolor", "biomechanical", "retro_futuristic",
"fighting_game", "mario", "pokemon", "street_fighter", "horror",
"manga", "space", "paper_mache", "tilt_shift".}
}
\value{
List of prompt and negative prompt
}
\description{
Get prompt and negative prompt based on style
}
