% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_audio.R
\name{get_audio}
\alias{get_audio}
\title{Get audio from text}
\usage{
get_audio(text)
}
\arguments{
\item{text}{Vector of story text}
}
\value{
List of audio in binary.
}
\description{
Get audio from text
}
