% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_single_image.R
\name{get_image}
\alias{get_image}
\title{Get image if request is successful}
\usage{
get_image(response)
}
\arguments{
\item{response}{Response from Workers AI Model API}
}
\value{
Image or NULL.
}
\description{
Get image if request is successful
}
