% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_images.R
\name{get_images}
\alias{get_images}
\title{Get all images corresponding to prompts}
\usage{
get_images(response_prompts, style)
}
\arguments{
\item{response_prompts}{A vector of prompts that describe what to draw}

\item{style}{A character string specifying the desired style. Choices are:
"anime", "comics", "lego_movie", "play_doh", "ethereal_fantasy", "isometric",
"line_art", "origami", "pixel_art", "abstract", "impressionist",
"renaissance", "watercolor", "biomechanical", "retro_futuristic",
"fighting_game", "mario", "pokemon", "street_fighter", "horror",
"manga", "space", "paper_mache", "tilt_shift".}
}
\value{
List of raw images.
}
\description{
Get all images corresponding to prompts
}
