% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prompts_for_images.R
\name{get_prompts_for_images}
\alias{get_prompts_for_images}
\title{Request story from API}
\usage{
get_prompts_for_images(
  story,
  max_tokens = 1000,
  ACCOUNT_ID = Sys.getenv("ACCOUNT_ID"),
  API_KEY = Sys.getenv("API_KEY"),
  base_url = cf_base_url()
)
}
\arguments{
\item{story}{Vector of story}

\item{max_tokens}{Max number of tokens}

\item{ACCOUNT_ID}{Cloudflare Workers AI Model API account ID}

\item{API_KEY}{Cloudflare Workers AI Model API key}

\item{base_url}{Base URL of Workers AI Model API}
}
\value{
Character vector with multiple sentences.
}
\description{
Request story from API
}
