% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_story.R
\name{get_story}
\alias{get_story}
\title{Request story from API}
\usage{
get_story(
  prompt,
  num_of_sentences = 5,
  max_tokens = 1000,
  ACCOUNT_ID = Sys.getenv("ACCOUNT_ID"),
  API_KEY = Sys.getenv("API_KEY"),
  base_url = cf_base_url()
)
}
\arguments{
\item{prompt}{Description of story}

\item{num_of_sentences}{Number of sentences in the story}

\item{max_tokens}{Max number of tokens}

\item{ACCOUNT_ID}{Cloudflare Workers AI Model API account ID}

\item{API_KEY}{Cloudflare Workers AI Model API key}

\item{base_url}{Base URL of Workers AI Model API}
}
\value{
Character vector with multiple sentences.
}
\description{
Request story from API
}
