% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_single_image.R
\name{req_single_image}
\alias{req_single_image}
\title{Request a single image from API}
\usage{
req_single_image(
  prompt,
  instructions,
  ACCOUNT_ID = Sys.getenv("ACCOUNT_ID"),
  API_KEY = Sys.getenv("API_KEY"),
  base_url = cf_base_url()
)
}
\arguments{
\item{prompt}{Description of image}

\item{instructions}{Instructions for image drawing}

\item{ACCOUNT_ID}{Cloudflare Workers AI Model API account ID}

\item{API_KEY}{Cloudflare Workers AI Model API key}

\item{base_url}{Base URL of Workers AI Model API}
}
\value{
Request.
}
\description{
Request a single image from API
}
